refactor: vote 카운트 타입을 long으로 통일

+++ src/main/java/com/umc/product/notice/application/port/in/query/dto/NoticeVoteInfo.java
@@ -17,7 +17,7 @@ public record NoticeVoteInfo(
     VoteStatus status,
     Instant startsAt,
     Instant endsAtExclusive,
-    int totalParticipants,
+    long totalParticipants,
     List<VoteOptionInfo> options,
     List<Long> mySelectedOptionIds) {
     /**
@@ -26,7 +26,7 @@ public record NoticeVoteInfo(
     public record VoteOptionInfo(
         Long optionId,
         String content,
-        int voteCount,
+        long voteCount,
         BigDecimal voteRate,
         List<Long> selectedMemberIds) {
     }
